refactor: extract shared agent ops and split executor main.rs
- Extract duplicated spawn_child_agent/check_children_status logic from
  temper.rs and local.rs into shared tools/agent_ops.rs module
- Add MAX_CHILDREN_BUDGET (50) cap on check_children_status iterations
  (TigerStyle: budgets not limits)
- Split temper-executor main.rs (645 lines) into health.rs, schedule.rs,
  agent_type.rs, and daemon.rs modules (main.rs now 331 lines)
- Resolve model from AgentType in schedule ticker instead of hardcoding
  "claude-sonnet-4-6"
