+++ crates/temper-executor/src/main.rs
@@ -17,19 +17,25 @@
 //! temper-executor --detach --health-port 4201
 //! ```
 
+mod agent_type;
+mod daemon;
+mod health;
+mod schedule;
+
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 
 use anyhow::{Context, Result};
-use axum::extract::State as AxumState;
-use axum::response::Json;
-use axum::routing::get;
 use clap::Parser;
 use futures_util::StreamExt;
-use temper_agent_runtime::{AgentRunner, AnthropicProvider, LocalToolRegistry, TemperToolRegistry};
 use temper_sdk::TemperClient;
 use tokio::sync::Semaphore;
 use tracing::{error, info, warn};
+
+use crate::agent_type::run_agent;
+use crate::daemon::{cleanup_pid_file, daemonize};
+use crate::health::{HealthState, run_health_server};
+use crate::schedule::run_schedule_ticker;
 
 /// Headless agent executor for Temper.
 ///
@@ -77,15 +83,6 @@ fn hostname() -> String {
     std::env::var("HOSTNAME") // determinism-ok: executor process, not simulation-visible
         .or_else(|_| std::env::var("HOST"))
         .unwrap_or_else(|_| "unknown".to_string())
-}
-
-/// Shared state for the health endpoint.
-#[derive(Clone)]
-struct HealthState {
-    executor_id: String,
-    max_concurrent: usize,
-    active_agents: Arc<AtomicUsize>,
-    shutting_down: Arc<AtomicBool>,
 }
 
 #[tokio::main]
@@ -200,31 +197,6 @@ async fn main() -> Result<()> {
     }
 
     info!("Executor shutdown complete");
-    Ok(())
-}
-
-/// Health check endpoint handler.
-async fn health_handler(AxumState(state): AxumState<HealthState>) -> Json<serde_json::Value> {
-    Json(serde_json::json!({
-        "status": if state.shutting_down.load(Ordering::Relaxed) { "draining" } else { "healthy" },
-        "executor_id": state.executor_id,
-        "active_agents": state.active_agents.load(Ordering::Relaxed),
-        "max_concurrent": state.max_concurrent,
-    }))
-}
-
-/// Run the health check HTTP server.
-async fn run_health_server(port: u16, state: HealthState) -> Result<()> {
-    let app = axum::Router::new()
-        .route("/health", get(health_handler))
-        .with_state(state);
-    let listener = tokio::net::TcpListener::bind(format!("0.0.0.0:{port}"))
-        .await
-        .context("Failed to bind health port")?;
-    info!(port = port, "Health endpoint listening");
-    axum::serve(listener, app)
-        .await
-        .context("Health server error")?;
     Ok(())
 }
 
@@ -356,290 +328,4 @@ async fn run_event_loop(
     }
 
     anyhow::bail!("SSE stream ended unexpectedly")
-}
-
-/// Default system prompt when no AgentType is configured.
-const DEFAULT_SYSTEM_PROMPT: &str = "You are a Temper agent. Accomplish your assigned goal \
-    using the tools available to you. Report results clearly.\n\n\
-    ## Delegation\n\
-    For complex tasks, you can delegate sub-tasks to child agents:\n\
-    - `spawn_child_agent(role, goal, model)` — spawns a child that runs autonomously\n\
-    - `check_children_status()` — check progress of all spawned children\n\
-    You cannot complete until all children have finished (Completed or Failed).";
-
-/// Resolve an AgentType entity for the given agent, returning (system_prompt, tool_set, model).
-///
-/// Falls back to CLI defaults when the agent has no agent_type_id or the AgentType
-/// entity is not found.
-async fn resolve_agent_type(
-    client: &TemperClient,
-    agent: &serde_json::Value,
-    default_tool_mode: &str,
-    default_model: &str,
-) -> (String, String, String) {
-    let agent_type_id = agent
-        .get("agent_type_id")
-        .or_else(|| agent.get("fields").and_then(|f| f.get("agent_type_id")))
-        .and_then(|v| v.as_str())
-        .unwrap_or("");
-
-    if agent_type_id.is_empty() {
-        return (
-            DEFAULT_SYSTEM_PROMPT.to_string(),
-            default_tool_mode.to_string(),
-            default_model.to_string(),
-        );
-    }
-
-    match client.get("AgentTypes", agent_type_id).await {
-        Ok(at) => {
-            let resolve = |key: &str, default: &str| -> String {
-                at.get(key)
-                    .or_else(|| at.get("fields").and_then(|f| f.get(key)))
-                    .and_then(|v| v.as_str())
-                    .filter(|s| !s.is_empty())
-                    .unwrap_or(default)
-                    .to_string()
-            };
-            let prompt = resolve("system_prompt", DEFAULT_SYSTEM_PROMPT);
-            let tool_set = resolve("tool_set", default_tool_mode);
-            let model = resolve("model", default_model);
-            info!(
-                agent_type_id = %agent_type_id,
-                model = %model,
-                tool_set = %tool_set,
-                "Resolved AgentType"
-            );
-            (prompt, tool_set, model)
-        }
-        Err(e) => {
-            warn!(
-                agent_type_id = %agent_type_id,
-                "Failed to resolve AgentType: {e}. Using defaults."
-            );
-            (
-                DEFAULT_SYSTEM_PROMPT.to_string(),
-                default_tool_mode.to_string(),
-                default_model.to_string(),
-            )
-        }
-    }
-}
-
-/// Run a single agent to completion.
-async fn run_agent(
-    temper_url: &str,
-    tenant: &str,
-    agent_id: &str,
-    tool_mode: &str,
-    model: &str,
-) -> Result<()> {
-    info!(agent_id = %agent_id, "Starting agent execution");
-
-    let client = TemperClient::new(temper_url, tenant);
-
-    // Fetch agent entity to resolve AgentType.
-    let agent = client.get("Agents", agent_id).await?;
-    let (system_prompt, resolved_tool_mode, resolved_model) =
-        resolve_agent_type(&client, &agent, tool_mode, model).await;
-
-    let provider = AnthropicProvider::new(&resolved_model)?;
-
-    let tools: Box<dyn temper_agent_runtime::ToolRegistry> = match resolved_tool_mode.as_str() {
-        "temper" => Box::new(TemperToolRegistry::new(TemperClient::new(
-            temper_url, tenant,
-        ))),
-        _ => Box::new(LocalToolRegistry::new(TemperClient::new(
-            temper_url, tenant,
-        ))),
-    };
-
-    let principal_id = std::sync::Arc::new(std::sync::Mutex::new(Some(agent_id.to_string())));
-    let runner = AgentRunner::new(client, Box::new(provider), tools, principal_id);
-    runner.resume(agent_id, &system_prompt).await?;
-
-    info!(agent_id = %agent_id, "Agent execution completed");
-    Ok(())
-}
-
-/// Schedule ticker: periodically evaluates Active schedules and fires due ones.
-///
-/// Runs on a 60-second interval. For each Active Schedule entity, parses the
-/// cron expression and fires `Schedule.Fire` if the cron is due. The Fire action's
-/// spawn effect creates an Agent entity, which the SSE event loop picks up.
-async fn run_schedule_ticker(temper_url: &str, tenant: &str, shutting_down: &Arc<AtomicBool>) {
-    use chrono::Utc; // determinism-ok: executor process, not simulation-visible
-    use cron::Schedule;
-    use std::str::FromStr;
-
-    let client = TemperClient::new(temper_url, tenant);
-    let mut interval = tokio::time::interval(std::time::Duration::from_secs(60));
-
-    loop {
-        interval.tick().await;
-        if shutting_down.load(Ordering::Relaxed) {
-            break;
-        }
-
-        // Query Active schedules.
-        let schedules = match client
-            .list_filtered("Schedules", "status eq 'Active'")
-            .await
-        {
-            Ok(s) => s,
-            Err(e) => {
-                warn!("Failed to query schedules: {e}");
-                continue;
-            }
-        };
-
-        let now = Utc::now(); // determinism-ok: executor process
-
-        for sched in schedules {
-            let field = |key: &str| -> &str {
-                sched
-                    .get(key)
-                    .or_else(|| {
-                        sched
-                            .get("fields")
-                            .and_then(|f: &serde_json::Value| f.get(key))
-                    })
-                    .and_then(|v: &serde_json::Value| v.as_str())
-                    .unwrap_or_default()
-            };
-            let sched_id = field("id");
-            let cron_expr = field("cron_expr");
-            let last_run = field("last_run");
-            let run_count: u64 = field("run_count").parse().unwrap_or(0);
-            let max_runs: u64 = field("max_runs").parse().unwrap_or(0);
-
-            // Check max_runs (0 = unlimited).
-            if max_runs > 0 && run_count >= max_runs {
-                // Auto-complete the schedule.
-                if let Err(e) = client
-                    .action("Schedules", sched_id, "Complete", serde_json::json!({}))
-                    .await
-                {
-                    warn!(schedule_id = %sched_id, "Failed to complete schedule: {e}");
-                }
-                continue;
-            }
-
-            // Parse cron expression.
-            let schedule = match Schedule::from_str(cron_expr) {
-                Ok(s) => s,
-                Err(e) => {
-                    warn!(schedule_id = %sched_id, cron = %cron_expr, "Invalid cron expression: {e}");
-                    continue;
-                }
-            };
-
-            // Check if due: find next occurrence after last_run (or epoch if never run).
-            let last = if last_run.is_empty() {
-                chrono::DateTime::<Utc>::MIN_UTC
-            } else {
-                last_run
-                    .parse::<chrono::DateTime<Utc>>()
-                    .unwrap_or(chrono::DateTime::<Utc>::MIN_UTC)
-            };
-
-            let next = schedule.after(&last).next();
-            let is_due = next.is_some_and(|n| n <= now);
-
-            if !is_due {
-                continue;
-            }
-
-            // Resolve agent params from the schedule entity.
-            let resolve = |key: &str| -> String {
-                sched
-                    .get(key)
-                    .or_else(|| sched.get("fields").and_then(|f| f.get(key)))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or_default()
-                    .to_string()
-            };
-
-            let agent_role = resolve("agent_role");
-            let goal_template = resolve("goal_template");
-            let agent_type_id = resolve("agent_type_id");
-            let now_str = now.to_rfc3339();
-
-            info!(
-                schedule_id = %sched_id,
-                role = %agent_role,
-                "Firing schedule"
-            );
-
-            if let Err(e) = client
-                .action(
-                    "Schedules",
-                    sched_id,
-                    "Fire",
-                    serde_json::json!({
-                        "last_run": now_str,
-                        "role": agent_role,
-                        "goal": goal_template,
-                        "model": "claude-sonnet-4-6",
-                        "agent_type_id": agent_type_id,
-                    }),
-                )
-                .await
-            {
-                warn!(schedule_id = %sched_id, "Failed to fire schedule: {e}");
-            }
-        }
-    }
-}
-
-/// Double-fork daemonization with PID file.
-fn daemonize() -> Result<()> {
-    use std::fs;
-    use std::os::unix::process::CommandExt;
-    use std::process::Command;
-
-    // Create PID file directory.
-    let pid_dir = dirs_pid_dir();
-    fs::create_dir_all(&pid_dir).context("Failed to create PID directory")?;
-
-    // Fork: the child continues, the parent exits.
-    // We use a re-exec approach instead of raw fork for safety.
-    let args: Vec<String> = std::env::args().filter(|a| a != "--detach").collect();
-    let child = Command::new(&args[0])
-        .args(&args[1..])
-        .stdin(std::process::Stdio::null())
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null())
-        .process_group(0)
-        .spawn()
-        .context("Failed to spawn daemon process")?;
-
-    // Write PID file.
-    let pid_file = pid_dir.join("executor.pid");
-    fs::write(&pid_file, child.id().to_string()).context("Failed to write PID file")?;
-    eprintln!(
-        "Executor daemonized. PID={}, PID file={}",
-        child.id(),
-        pid_file.display()
-    );
-
-    // The parent exits immediately.
-    std::process::exit(0);
-}
-
-/// Clean up PID file on shutdown.
-fn cleanup_pid_file() {
-    let pid_file = dirs_pid_dir().join("executor.pid");
-    if pid_file.exists() {
-        std::fs::remove_file(&pid_file).ok();
-    }
-}
-
-/// PID file directory: ~/.local/state/temper/
-fn dirs_pid_dir() -> std::path::PathBuf {
-    let home = std::env::var("HOME").unwrap_or_else(|_| "/tmp".to_string()); // determinism-ok: executor process
-    std::path::PathBuf::from(home)
-        .join(".local")
-        .join("state")
-        .join("temper")
 }
